+++ packages/react-universal/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+* Use snake cased `history.to_href`
+
 ## 2.0.0-alpha.2
 
 * Remove `useBlock` hook